Improve option decorator definitions
When there is more the one click option in a reusable decorator, it can
be difficult to follow and detect ommitting to call the `fn`. Also, the
order of options in the source code is the reverse of how they appear in
the help text. Fix these issues by moving to a list-based pattern when
there is more than one option in a single decorator.

Also, simplify expressions to drop unnecessary assignments.