Remove `account lntn-balances` command
This command was fine in an early testnet environment, with a handful of
stable validators.

However, the command no longer provides comprehensive results, as it
cannot return LNTN balances for validators that are no longer in the
committee. So if the goal is to help the user track their balances, the
implementation is of limited value. Also, as a correct solution would
require an off-chain indexer, it is it is outside the scope of this
tool.

Also, the implementation is extremely slow and can take minutes to
finish iterating all the current committee members. This problem will
only get worse as committee sizes increase.