Drop the --keyfile option for signing
This feature is being removed because reading a plaintext password
from a command argument is one the most insecure methods of reading
a password.

If you need to automate signing with a keyfile, you may continue to
use the environment variable `KEYFILEPWD`, which will be checked first
before prompting for a keyfile password.