Drop the --password option for signing
This feature is removed because it is highly insecure. Reading a
plaintext password from a command argument is expose the password in
plaintext in  shell history.

For automated signing using a keyfile, the environment variable
`KEYFILEPWD` can be used, which will be checked first, before falling
back to prompting for keyfile password.