Fix auth options should not be required
In several commands, such as `account balance`, the account to use
can be specified as an argument. And in such cases, the current
`RequiredAnyOptionGroup` will raise an error that an auth option was
not provided.

To fix this, the decorator can be parameterized with a
`required` boolean, but this is not even necessary, because the
`validate_authenticator()` and `validate_authenticator_account()`
already perform sanity checks, so best to push the logic back to there.