Add check for address mismatch when signing tx
With keyfiles, we deferred to eth-account's logic to check for mismatch
between the TX from-address and the identity of the signer in `account
signtx` and `tx sign`. With the new authenticator protocol, these checks
should be explicit as trezorlib doesn't accept a `from` field, and so
any from-address set in the TX would be ignored at signing time.

Ideally, Autonity CLI commands should not accept a --from address, and
should not create TX with a `from` address, as this field is not part of
signed transaction data.